requirements.txt
asgiref==3.8.1
attrs==23.2.0
autobahn==23.6.2
Automat==22.10.0
cffi==1.16.0
channels==4.0.0
colorama==0.4.6
constantly==23.10.4
cryptography==42.0.7
daphne==4.1.2
Django==4.1.13
django-cors-headers==4.3.1
djangorestframework==3.15.1
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.3.1
djongo==1.3.6
dnspython==2.6.1
gunicorn==22.0.0
hyperlink==21.0.0
idna==3.7
incremental==22.10.0
iniconfig==2.0.0
packaging==24.0
pluggy==1.5.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
PyJWT==1.7.1
pymongo==3.12.3
pyOpenSSL==24.1.0
pytest==8.2.2
pytz==2024.1
service-identity==24.1.0
six==1.16.0
sqlparse==0.2.4
Twisted==24.3.0
twisted-iocpsupport==1.0.4
txaio==23.1.1
typing_extensions==4.12.0
tzdata==2024.1
zope.interface==6.4.post2
